requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
pyodbc==5.1.0
Flask-Mail==0.9.1
APScheduler==3.10.4
Werkzeug==3.0.3
requests==2.32.3